% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{genData}
\alias{genData}
\title{Generate synthetic data martix}
\usage{
genData(
  n_features = 200,
  n_samples = 50,
  n_factors = 5,
  theta_z = 1,
  theta_w = 1,
  tau = 0,
  tau_mode = "perProt",
  shape_noise = 1.5,
  scale_noise = 1.5,
  alpha_row = 0,
  sd_row = 1,
  alpha_col = 20,
  sd_col = 2
)
}
\arguments{
\item{n_features}{a number of features}

\item{n_samples}{a number of samples/observations}

\item{n_factors}{a number of factors}

\item{theta_z}{a probability of factor actication}

\item{theta_w}{a probability of weight actication}

\item{tau}{the precision for likelihood}

\item{tau_mode}{a character that can take the value of "perFeature" or None.
"perFeature" allows the noise to varied to degrees defined by "shape_noise" and "scale_noise" in a per-feature manner or by reference_df.}

\item{shape_noise}{the shape parameter of gamma distribution}

\item{scale_noise}{the scale parameter of gamma distribution}

\item{alpha_row}{intercepts are added to row level using a normal distribution. alpha_row is the loc parameter of this distribution.}

\item{sd_row}{intercepts are added to row level using a normal distribution. sd_row is the scale parameter of this distribution.}

\item{alpha_col}{intercepts are added to column level using a normal distribution. alpha_col is the loc parameter of this distribution.}

\item{sd_col}{intercepts are added to column level using a normal distribution. sd_col is the loc parameter of this distribution.}
}
\value{
dataList
}
\description{
This function create complete data through a multiplication by a synthetic weigh matrix and factor matrix
}
