% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{genProbMiss}
\alias{genProbMiss}
\title{Generate synthetic data martix}
\usage{
genProbMiss(
  X,
  rho = 20,
  zeta = 2,
  model = "perFeature",
  rho_sd = 1,
  zeta_sd = 1,
  subSample = 0,
  filter_threshold = 0
)
}
\arguments{
\item{X}{a complete input matrix, with features in rows and samples in columns}

\item{rho}{loc parameter for dropout model}

\item{zeta}{slope parameter for the drop-out model}

\item{model}{a character that can take the value of "global", "perFeature" or "perSample". "global" indicates generating missing values for all measurements using the same probabilistic.
"perFeature" and "perSample" allow the drop-out model parameter to varied to degrees defined by "rho_sd" and "zeta_sd" in a per-feature or per-sample manner.}

\item{rho_sd}{the degree (standard deviation in a normal distribution) the loc of the drop-out model can vary.}

\item{zeta_sd}{the degree (standard deviation in a normal distribution) the slope of the drop-out model can vary.}

\item{subSample}{if value is > 0 and < 1, features will be subsetted randomly by this percentage.}

\item{filter_threshold}{the cutoff for removing proteins within missingness}
}
\value{
dataList
}
\description{
This function create complete data through a multiplication by a synthetic weigh matrix and factor matrix
}
