% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getParams}
\alias{getParams}
\title{Extract all parameters from the trained msBayes_model}
\usage{
getParams(msBayes_model, miss_data)
}
\arguments{
\item{msBayes_model}{a msBayesImpute object}

\item{miss_data}{the data that is used for training msBayesImpute}
}
\value{
params
}
\description{
This function output all parameters trained in msBayesImpute model
}
