% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{runMsBayesImpute}
\alias{runMsBayesImpute}
\title{Train a msBayesImpute model and impute missing values}
\usage{
runMsBayesImpute(
  miss_data,
  n_components = NULL,
  convergence_mode = "fast",
  drop_factor_threshold = 0.01
)
}
\arguments{
\item{miss_data}{the data should filter out complete missing proteins and normally distributed after e.g. logarithm transformation}

\item{n_components}{user can fix the number of components on their own; By default, the model can automatically filter out the insignificant factors}

\item{convergence_mode}{two convergence options: fast by default, and slow}

\item{drop_factor_threshold}{a threshold to filter out significant latent factors;By default, the values is 0.01}
}
\value{
a list of a model and a data matrix
}
\description{
This function is called when the input data is prepared
}
